update login flow
