+++ app/src/main/java/com/ecosystem/kin/app/MainActivity.java
@@ -34,7 +34,9 @@ import com.kin.ecosystem.common.model.NativeOffer;
 import com.kin.ecosystem.common.model.NativeSpendOfferBuilder;
 import com.kin.ecosystem.common.model.OrderConfirmation;
 import com.kin.ecosystem.common.model.UserStats;
+import com.kin.ecosystem.common.model.WhitelistData;
 import java.util.Random;
+
 
 public class MainActivity extends AppCompatActivity {
 
@@ -151,6 +153,7 @@ public class MainActivity extends AppCompatActivity {
 			}
 		});
 
+		login();
 		final TextView userIdTextView = findViewById(R.id.user_id_textview);
 		userIdTextView.setText(getString(R.string.user_id, userID));
 		userIdTextView.setOnClickListener(new OnClickListener() {
@@ -169,6 +172,43 @@ public class MainActivity extends AppCompatActivity {
 			.setText(getString(R.string.version_name, BuildConfig.VERSION_NAME));
 		addNativeOffer();
 		addNativeOfferClickedObserver();
+	}
+
+	private void login() {
+		if (BuildConfig.IS_JWT_REGISTRATION) {
+			/**
+			 * SignInData should be created with registration JWT {see https://jwt.io/} created securely by server side
+			 * In the the this example {@link SignInRepo#getJWT} generate the JWT locally.
+			 * DO NOT!!!! use this approach in your real app.
+			 * */
+			String jwt = SignInRepo.getJWT(this);
+
+			Kin.login(jwt, new KinCallback<Void>() {
+				@Override
+				public void onResponse(Void response) {
+					Log.d(TAG, "JWT onResponse: login");
+				}
+
+				@Override
+				public void onFailure(KinEcosystemException exception) {
+					Log.e(TAG, "JWT onFailure: " + exception.getMessage());
+				}
+			});
+		} else {
+			/** Use {@link WhitelistData} for small scale testing */
+			WhitelistData whitelistData = SignInRepo.getWhitelistSignInData(this, getAppId(), getApiKey());
+			Kin.login(whitelistData, new KinCallback<Void>() {
+				@Override
+				public void onResponse(Void response) {
+					Log.d(TAG, "WhiteList onResponse: login");
+				}
+
+				@Override
+				public void onFailure(KinEcosystemException exception) {
+					Log.e(TAG, "WhiteList onFailure: " + exception.getMessage());
+				}
+			});
+		}
 	}
 
 	private void addNativeOffer() {
@@ -576,6 +616,16 @@ public class MainActivity extends AppCompatActivity {
 				.setTextColor(Color.RED);
 		}
 		snackbar.show();
+	}
+
+	@NonNull
+	public static String getAppId() {
+		return BuildConfig.SAMPLE_APP_ID;
+	}
+
+	@NonNull
+	public static String getApiKey() {
+		return BuildConfig.SAMPLE_API_KEY;
 	}
 
 	@Override
